refac: refactor code for cyclic dependency and test code
