+++ .gitignore
@@ -14,4 +14,5 @@ deploy_*.sh
 *.ini
 
 !pytest.ini
+!test_config.ini
 test.db
