+++ test_config.ini
@@ -1,0 +1,20 @@
+[SESSION]
+SECRET_KEY = test_session
+
+[JWT]
+ALGORITHM = HS256
+JWT_SECRET_KEY = test_jwt_secret_key
+JWT_REFRESH_SECRET_KEY = test_jwt_secret_key
+ACCESS_TOKEN_EXPIRE_MINUTES = 43200
+REFRESH_TOKEN_EXPIRE_MINUTES = 129600
+
+[AWS]
+AWS_ACCESS_KEY_ID = test_access_key_id
+AWS_SECRET_ACCESS_KEY = test_aws_secret_key
+REGION_NAME = test_region_name
+
+[S3]
+BUCKET = test_bucket
+
+[GOOGLE]
+CLIENT_ID = test_client_id
